requirements.txt
pandas==1.0.3
streamlit==0.73.1
SQLAlchemy==1.3.18
plotly==4.14.1
requests==2.23.0
azure-identity
azure-keyvault